+++ rust/origen/src/core/model/registers/bit_collection.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, Clone)]
+pub struct BitCollection {}
